+++ python/paddle/fluid/tests/unittests/test_elementwise_mul_op.py
@@ -391,8 +391,8 @@ class TestElementwiseMulop(unittest.TestCase):
     def func_dygraph_mul(self):
         paddle.disable_static()
 
-        np_a = np.random.random((2, 3, 4)).astype(np.float64)
-        np_b = np.random.random((2, 3, 4)).astype(np.float64)
+        np_a = np.random.random((2, 3, 4)).astype(np.float32)
+        np_b = np.random.random((2, 3, 4)).astype(np.float32)
 
         tensor_a = paddle.to_tensor(np_a, dtype="float32")
         tensor_b = paddle.to_tensor(np_b, dtype="float32")
